Framework: Forward port Power menu reboot by DvTonder
This is a stripped down version of his power menu config patch

Change-Id: I5cc9f900bdb2bb1f0a11014a09f8c792cac850e7

This commit responds to a setting in Development settings for including
options in the power menu for rebooting into recovery or bootloader. It
is defauled to off.

When enabled, the Advanced reboot options will only be available once the
device is unlocked.

Change-Id: I675f31de1f82c344c7923e953fd4a6e26dd036da