Keyguard: PIN lock input views flash when SIM status ready
When airplane mode is turned off and DSDS mode is on, cat service
will request RIL to update the SIM card status. When the broadcast
of SIM ready status is received, keyguardview will be reset. So the
PIN lock input view will flash and bring bad user experience.

As there is nothing to update by resetting keyguardview while SIM
is ready. So dismiss the ready status of SIM card.

CRs-Fixed: 537172

Change-Id: I510e8f1e2f08cb946fe47f02b015ae5bf89d5025