SystemUI: QuickTiles: add support for gestures (WIP)
allows to do actions on left and/or right swipe on a tile via
overriding onFlingLeft and onFlingRight.

for that we are implementing a gesture detector, which also handles
click and longpress events.

as example, on swipe right at the preferences tile, we launch
Device Control, on swipe left we launch Update Center.

Change-Id: I9580ba3a94d5d30859f3ac0a87d3564f971877be