SystemUI: QuickTiles: add support for gestures
allows to do actions on left and/or right swipe on a tile via
overriding onFlingLeft and onFlingRight.

for that we are implementing a gesture detector, which also handles
click and longpress events.

already make use of it at the preferences tile.
on swipe right we launch Device Control,
on swipe left we launch Update Center.

Change-Id: I9580ba3a94d5d30859f3ac0a87d3564f971877be