WallpaperCropper: Fix crash when set .gif as wallpaper
It returns an equivalent scaled cropBounds as true
crop bounds, then it set the rounded rect by rounding
"out" the true crop rect, this make the rounded crop
rect may out range of the original bitmap, then an error
will happen.

We change to check the rounded crop rect, if it is out
the range of original bitmap, we set the original bitmap
bounds as rounded crop bounds.

Change-Id: I91fee6e0413ecb46e71337cdc6c467e9915e3ce7
CRs-Fixed: 622582