Base: rewrite ScreenRecord service
we dont pass around data, actually we just do 3 things
    1) starting it
    2) stopping it
    3) toggling the pointer
so lets drop all the service connections and make it simple

everything is handled by the service itself, this removes all the
checks if the service is running or not.

add the action to start the screen record to our
action processor.

some parts are taken from omnirom's implementation

Change-Id: I055da4c359fb7be9a024e6a1ceabfc0310e7f906