Keyguard: Distinguish the TextView id of pinEntry
Both KeyguardSimPinView and KeyguardPinView have the same id of pinEntry
for their TextView of the password. When one of them is created in the
mSecurityViewContainer, the other one will get the wrong TextView of the
same id pinEntry.

Change the id pinEntry in KeyguardSimPinView to simPinEntry.

This fixes CYAN-3822,CYAN-3887

Change-Id: Ie78ca9a23498577ddde4458e041241bcedf49a8b
CRs-Fixed: 661955