Fix initial black lockscreen
if see through is enabled and no custom lockscreen wallpaper is set, the
lockscreen wallpapaer is black when unlocking the phone for the first time.
this looks ugly and irritates users.

do not pass false to updateShowWallpaper, rather check if we should show
the wallpaper with mKeyguardHost.shouldShowWallpaper() to prevent this issue

Change-Id: I07e40017cfeeb7fef83117cdebeb74707c232968