Recents long press: add pluggable implementation
Adds a new intent ACTION_RECENTS_LONG_PRESS for the
long press of the recents button. Other system applications
can register for the intent to provide custom handling.
Handlers of this intent are queried for in the settings
application, and if no handlers are found, the default
implementation falls back on ActionUtils.switchToLastApp().

The long press listener on the recents button is replaced
with a touch handler to allow for more custom handling.

Change-Id: I8a27ae0fe58d92a1626f0bd661c5f57c589fac17