Barrier only messages shouldn't prevent the idle handlers from running
If the queue just consists of barrier messages the idle handlers will
not been executed, cause mMessages isn't null. Barrier messages aren't
real messages and shouldn't be included in this test. They are an
internal feature of the message queue and they just happen to be
implemented as normal messages without a target.

This fix allows the idle handlers to run even if the queue isn't empty,
but only consists of barrier messages.

The bug can be seen if somebody calls Instrumentation.waitForIdleSync()
and the application changes the layout on every frame, like in
android.app.cts.ProgressDialogTest -- testShow3 on certain platforms.
The progress bar in indeterminate mode repaints the spinning circle on
every frame. This seems to trigger scheduleTraversals() in ViewRootImpl,
which itself inserts a barrier into the message queue.

Other people noticed this as well as seen here:
https://code.google.com/p/android/issues/detail?id=82046

I could repo this using the test case from the bug report on a Nexus 5.
Changing the testShow3 CTS test so that there is a Thread.sleep(3000);
right before mInstrumentation.waitForIdleSync(); also reproduce it on a
Nexus 5.

It is also worth to mention that this really depends on the type of
animation. Removing the minimum sdk version from the test application,
switches the animation to the old version of the spinning circle. This
seems to include lesser frames and therefore doesn't trigger this bug.

Change-Id: I193c955dde0210eca89b2286f8daf0c7ce06fa3b
Signed-off-by: Christian Poetzsch <christian.potzsch@imgtec.com>