SystemUI: fix recents task view height calculation
Initialize search bar height early enough for layouts to catch it and
not cause jumpiness in certain situations. For instance on a device with
a soft navigation bar, hitting recents from any application, right after
the animation would finish, it would relize its layout was incorrect and
relayout the thumbnails causing a jump.

Change-Id: I31a50f97e3799727a38b55beea1f003514331729